+++ start/src/App.vue
@@ -1,22 +1,5 @@
-<script lang="ts" setup>
-  import { ref } from 'vue'
+<script lang="ts" setup></script>
 
-  const current = ref(1)
-  const total = ref(100000)
+<template></template>
 
-  const pagesize = ref(10)
-
-  const pageSizes = ref([10, 20, 30, 40, 50])
-</script>
-
-<template>
-  <f-pagination
-    v-model:current="current"
-    v-model:page-size="pagesize"
-    round
-    jump-search
-    background
-    :total="total"
-    :page-sizes="pageSizes"
-  />
-</template>
+<style lang="scss" scoped></style>